simplifies coordinate storage and manipulation

+++ src/utils/intersects.rs
@@ -9,7 +9,7 @@ pub trait Intersects<T> {
     fn intersects(&self, val: &T) -> bool;
 }
 
-impl Intersects<Range<u64>> for Range<u64> {
+impl Intersects<Range<u32>> for Range<u32> {
     /// Check if two ranges intersect (overlap)
     ///
     /// ```
@@ -17,9 +17,11 @@ impl Intersects<Range<u64>> for Range<u64> {
     /// assert!((0..3).intersects(&(0..1)));
     /// assert!(!(0..3).intersects(&(5..7)));
     /// assert!(!(0..3).intersects(&(1..1)));
+    /// assert!(!(1..1).intersects(&(1..1)));
+    /// assert!(!(1..1).intersects(&(0..3)));
     /// assert!((1..3).intersects(&(0..2)));
     /// ```
-    fn intersects(&self, val: &Range<u64>) -> bool {
+    fn intersects(&self, val: &Range<u32>) -> bool {
         (self.start < val.end && self.end > val.start) && !(self.is_empty() || val.is_empty())
     }
 }
